e-run-j-run_task/ - RenderCanvas 생성 - RenderCanvas에서 canvas 관련 랜더링 메소드 동작
